Move TemplateLibrary to inside osm module
