+++ src/main/java/org/opentripplanner/openstreetmap/TemplateLibrary.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.graph_builder.module.osm;
+package org.opentripplanner.openstreetmap;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -6,6 +6,9 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import org.opentripplanner.openstreetmap.model.OSMWithTags;
 
+/**
+ * Create and localize strings based on OpenStreetMap tags and templates
+ */
 public class TemplateLibrary {
 
   private static final Pattern patternMatcher = Pattern.compile("\\{(.*?)}");
